fix: fix showing unwanted rotations in header back btn on state changes
